requirements.txt
Flask
python-dotenv
google-generativeai
requests
beautifulsoup4